refactor: simplify business page structure and enhance service list handling
- Refactored the BusinessPage component to utilize hooks for data fetching, improving readability and maintainability.
- Replaced the previous async data fetching method with a synchronous approach using `useBusiness` hook.
- Introduced a new `BusinessContent` component to encapsulate the business details rendering logic.
- Updated the ServiceList component to handle optional services and display a message when no services are available.
- Adjusted TypeScript types for better clarity and flexibility in service data handling.

Changes:
- Modified src/app/business/[slug]/page.tsx to streamline data fetching and rendering.
- Enhanced src/components/services/service-list.tsx to improve service availability handling.

This commit enhances the overall structure and user experience of the business page by leveraging React hooks and improving service list management.
