+++ src/app/business/[slug]/business-skeleton.tsx
@@ -1,0 +1,20 @@
+export function BusinessSkeletonLoader() {
+  return (
+    <div className="min-h-screen bg-gray-50 animate-pulse">
+      {/* Hero Section Skeleton */}
+      <div className="relative h-[480px] bg-gray-200" />
+      
+      {/* Content Skeleton */}
+      <div className="container mx-auto px-4 py-8 space-y-8">
+        <div className="bg-white rounded-2xl p-6">
+          <div className="h-8 bg-gray-200 rounded w-1/3 mb-4" />
+          <div className="grid grid-cols-1 sm:grid-cols-3 gap-4">
+            {[...Array(3)].map((_, i) => (
+              <div key={i} className="h-24 bg-gray-200 rounded-xl" />
+            ))}
+          </div>
+        </div>
+      </div>
+    </div>
+  )
+}
