+++ src/lib/get-query-client.ts
@@ -1,0 +1,4 @@
+import { QueryClient } from '@tanstack/react-query'
+import { cache } from 'react'
+
+export const getQueryClient = cache(() => new QueryClient())
